% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smithscrapr.R
\name{bio_track_matcher}
\alias{bio_track_matcher}
\title{Make tables for Biology's major requirements}
\usage{
bio_track_matcher(track)
}
\arguments{
\item{number}{A numeric value 1-5 that represent which track of biology the user is looking for}
}
\value{
A data frame that categorizes courses for major requirements under must or choose to take columns

Note that for each row, courses appears in one column
}
\description{
Given a list of web-scraped major requirements for each five of Biology's tracks, turn it into a cleaned data frame
}
\examples{
track_1 <- bio_track_matcher(1)
track_1

}
