% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sds_df}
\alias{sds_df}
\title{SDS Major requirement}
\format{
\describe{
\item{Requirement}{Requirement Categories for SDS}
\item{Must}{Courses required to take for SDS major}
\item{Choose}{Courses to choose to take or not for SDS major}
}
}
\usage{
sds_df
}
\description{
SDS Major requirement
}
\keyword{datasets}
