% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smithscrapr.R
\name{req_df}
\alias{req_df}
\title{Make tables for major requirements}
\usage{
req_df(major)
}
\arguments{
\item{character}{A character vector of a major's abbreviation from below: "sds", "csc", "eco",
"bch", "ast", or "chm"}
}
\value{
A data frame that categorizes courses for major requirements under must or choose to
take columns

Note that for each row, courses appears in one column
}
\description{
Given a list of web-scraped major requirements, turn it into a cleaned data frame
}
\examples{
sds_df <- req_df("sds")
sds_df

}
